+++ lib/shakapacker/webpack_runner.rb
@@ -6,7 +6,7 @@ module Shakapacker
   class WebpackRunner < Shakapacker::Runner
     def self.run(argv)
       $stdout.sync = true
-      ENV["NODE_ENV"] ||= (ENV["RAILS_ENV"] == "production") ? "production" : "development"
+      Shakapacker.ensure_node_env!
       new(argv).run
     end